feat: use external index for standalone testbed
