+++ testbed-create.py
@@ -1,14 +1,13 @@
 #!/usr/bin/env python3
-from datetime import datetime
+from utils import load_index
 import json
+import itertools
 import argparse
-import os
-
 
 if __name__ == "__main__":
   parser = argparse.ArgumentParser()
-  parser.add_argument('manifest',
-    help="package index manifest (e.g., from query)")
+  parser.add_argument('index',
+    help="package index (directory or manifest)")
   parser.add_argument('toolchain',
     help="Lean toolchain on build the packages on")
   parser.add_argument('-n', '--num', type=int, default=10,
@@ -24,9 +23,6 @@ if __name__ == "__main__":
     with open(file, 'r') as f:
       for line in f: exclusions.add(line.strip())
 
-  with open(args.manifest, 'r') as f:
-    pkgs = json.load(f)
-
   def create_entry(pkg: 'dict[str, any]'):
     return {
       'id': pkg['id'],
@@ -35,12 +31,13 @@ if __name__ == "__main__":
       'toolchain': args.toolchain
     }
 
+  pkgs = load_index(args.index)
   pkgs = filter(lambda pkg: pkg['fullName'] not in exclusions, pkgs)
   pkgs = map(create_entry, pkgs)
-  pkgs = list(pkgs)[0:args.num]
+  pkgs = list(itertools.islice(pkgs, args.num))
 
-  if args.output is not None:
-    print(json.dumps(pkgs, indent=2))
+  if args.output is None:
+    print(json.dumps(pkgs))
   else:
     with open(args.output, 'w') as f:
-      f.write(json.dumps(pkgs, indent=2))
+      f.write(json.dumps(pkgs))
